Have a query language by merging models
